Add control panel for volume control
You can now control the volume for each application by a trackbar. By clicking on the volume label on the right side a lock will appear and it will lock the volume to the current one. If other applications try to change the volume, ControlMic will change it back to the locked value